requirements.txt
ollama>=0.1.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.3.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0
mcp>=1.0.0
fastmcp>=0.9.0
pydantic>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
duckduckgo-search>=4.1.0
yt-dlp>=2023.12.0
requests>=2.31.0
gdown>=4.7.0
